v1.1.4

+++ yttelegram-buttons/src/main/java/cs/youtrade/telegram/buttons/data/AbstractUserData.java
@@ -13,7 +13,7 @@ public abstract class AbstractUserData {
     @Setter
     private Class<?> supportedEdit;
     @Setter
-    private long lastMessageId = -1L;
+    private int lastMessageId = -1;
 
     @Override
     public boolean equals(Object o) {
